Extract some utilities into a separate package
