+++ packages/react-dev-utils/checkRequiredFiles.js
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+var fs = require('fs');
+var path = require('path');
+var chalk = require('chalk');
+
+function checkRequiredFiles(files) {
+  var currentFilePath;
+  try {
+    files.forEach(filePath => {
+      currentFilePath = filePath;
+      fs.accessSync(filePath, fs.F_OK);
+    });
+    return true;
+  } catch (err) {
+    var dirName = path.dirname(currentFilePath);
+    var fileName = path.basename(currentFilePath);
+    console.log(chalk.red('Could not find a required file.'));
+    console.log(chalk.red('  Name: ') + chalk.cyan(fileName));
+    console.log(chalk.red('  Searched in: ') + chalk.cyan(dirName));
+    return false;
+  }
+}
+
+module.exports = checkRequiredFiles;
