+++ packages/react-dev-utils/openChrome.applescript
@@ -1,4 +1,3 @@
--- @remove-on-eject-begin
 (*
 Copyright (c) 2015-present, Facebook, Inc.
 All rights reserved.
@@ -7,7 +6,6 @@ This source code is licensed under the BSD-style license found in the
 -- LICENSE file in the root directory of this source tree. An additional grant
 of patent rights can be found in the PATENTS file in the same directory.
 *)
--- @remove-on-eject-end
 
 on run argv
   set theURL to item 1 of argv
