+++ packages/react-scripts/utils/createJestConfig.js
@@ -1,4 +1,3 @@
-// @remove-on-eject-begin
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
  * All rights reserved.
@@ -7,10 +6,11 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  */
-// @remove-on-eject-end
+
+// Note: this file does not exist after ejecting.
 
 const pathExists = require('path-exists');
-const paths = require('../../config/paths');
+const paths = require('../config/paths');
 
 module.exports = (resolve, rootDir, isEjecting) => {
   const setupFiles = [resolve('config/polyfills.js')];
